Created a function that retrieves emissions data
